+++ .circleci/example-refresh-project/green/green.clj
@@ -1,0 +1,6 @@
+(ns green
+  (:require
+   [red]))
+
+(defn ok []
+  42)