+++ .circleci/example-refresh-project/red/red.clj
@@ -1,0 +1,4 @@
+(ns red)
+
+(defn reflection [x]
+  (.method x))